demo/requirements.txt
Flask>=2.3.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
Werkzeug>=2.3.0
numpy>=1.24.0
